move code around and comment
