+++ service/controller/resource/etcd-certificates/doc.go
@@ -1,0 +1,4 @@
+package etcdcertificates
+
+// Package etcdcertificates implements a resource in charge of finding
+// necessary TLS certificates to communicate with Etcd endpoints.
