+++ service/controller/resource/etcd-certificates/resource.go
@@ -1,7 +1,6 @@
 package etcdcertificates
 
 import (
-	"context"
 	"reflect"
 
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
@@ -11,7 +10,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/generic"
-	"github.com/giantswarm/prometheus-meta-operator/service/key"
 )
 
 const (
@@ -23,8 +21,7 @@ type Config struct {
 	Logger    micrologger.Logger
 }
 
-// secretCopier provides a `ToCR` method which copies data from the source
-// cluster secret CR
+// secretCopier provides a way to create a new secret from different data source.
 type secretCopier struct {
 	logger     micrologger.Logger
 	clientFunc func(string) generic.Interface
@@ -39,6 +36,7 @@ func New(config Config) (*generic.Resource, error) {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
 	}
 
+	// Wrapping the secret client into a generic interface.
 	clientFunc := func(namespace string) generic.Interface {
 		c := config.K8sClient.K8sClient().CoreV1().Secrets(namespace)
 		return wrappedClient{client: c}
@@ -55,7 +53,7 @@ func New(config Config) (*generic.Resource, error) {
 		Logger:           config.Logger,
 		Name:             Name,
 		GetObjectMeta:    getObjectMeta,
-		GetDesiredObject: sc.ToCR,
+		GetDesiredObject: sc.ToSecret,
 		HasChangedFunc:   hasChanged,
 	}
 	r, err := generic.New(c)
@@ -66,37 +64,7 @@ func New(config Config) (*generic.Resource, error) {
 	return r, nil
 }
 
-func getObjectMeta(ctx context.Context, v interface{}) (metav1.ObjectMeta, error) {
-	cluster, err := key.ToCluster(v)
-	if err != nil {
-		return metav1.ObjectMeta{}, microerror.Mask(err)
-	}
-
-	return metav1.ObjectMeta{
-		Name:      key.EtcdSecret(v),
-		Namespace: key.Namespace(cluster),
-	}, nil
-}
-
-func (sc *secretCopier) ToCR(ctx context.Context, v interface{}) (metav1.Object, error) {
-	objectMeta, err := getObjectMeta(ctx, v)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	data, err := sc.getSource(ctx, v)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	secret := &corev1.Secret{
-		ObjectMeta: objectMeta,
-		StringData: data,
-	}
-
-	return secret, nil
-}
-
+// hasChanged determines if secret data have changed.
 func hasChanged(current, desired metav1.Object) bool {
 	c := current.(*corev1.Secret)
 	d := desired.(*corev1.Secret)
